+++ alu-web/src/teachers/index.ts
@@ -1,11 +1,9 @@
-import { ClassroomsHomepage } from './home';
 import { ClassroomDetail } from './detail';
 import { ClassroomStudentDetail } from './student';
 import { StudyAssignment } from './study-assignment';
 
 
 export {
-  ClassroomsHomepage,
   ClassroomDetail,
   ClassroomStudentDetail,
   StudyAssignment,