+++ teachers/urls.py
@@ -1,10 +1,12 @@
-from utils.utils import render_basic_view
 from django.urls import path
 
-from .views import *
+from .views import (
+    classroom_detail,
+    classroom_student_detail,
+    classroom_assignment_study,
+)
 
 urlpatterns = [
-    path('home/classrooms/', render_basic_view('teachers/home.html')),
     path('classrooms/<int:classroom_id>/', classroom_detail),
     path('classrooms/<int:classroom_id>/student/', classroom_student_detail),
     path('classrooms/<int:classroom_id>/assignments/<int:assignment_id>/study/', classroom_assignment_study),