+++ pages/markdown/misc/changelog.md
@@ -11,6 +11,7 @@ Many teachers like to create two versions of every assignment: one with all of t
 * **Fixed Older Notifications:** Fixed the link to older notifications and fixed the page that displayed older notifications.  I'm aware the notification system still isn't perfect, but it currently isn't used enough to justify committing serious time to improve it.
 * **Reversed Flashcard Browse Order:** Reversed the default flashcard browse order so that most recent flashcards are displayed first.  You can still reverse this order by clicking "Sort Ascending" at the top of the page.
 * **Fixed Flashcard Rearranging:** Flashcard rearranging was broken when the browsing flashcard list was reversed.  I've fixed this now.
+* **Removed Old Classroom Homepage:** Removed the old, deprecated classroom homepage from before Alu 0.5.0.
 
 
 ## Automatic Copied Deck Updates - Mar 10 - 0.5.3